profile.users and connections imported
there where many import errors, mainly caused by circular imports. to avoid circular imports, there are no more shortcuts in __init__.py files. its discussable, but i think, its better not to do that and force users to import explicit.
