--- facebook/modules/profile/__init__.py
@@ -1,1 +1,0 @@
-from .models import Profile
